==> Youtube-Video-Classification-on-Twitter/scripts/1-batch_mongoDB.ipynb <==
# Database connection

from pymongo import MongoClient
import json
import re
import csv
import pandas as pd
import json
from bson import ObjectId

# Create a MongoClient object and specify the MongoDB connection URL
url = "mongodb://localhost:27017/"
client1 = MongoClient(url)

# Access a specific database
db1 = client1["youtube_twitter_db1"]

# Send a ping to confirm a successful connection
try:
    client1.admin.command('ping')
    print("Pinged your deployment. You successfully connected to MongoDB!")
except Exception as e:
    print(e)

# Twitter collection creation

# Reading CSV and collection creation
tw_collection = db1.tw_collection
i = 0

with open('./df_youtube.csv', 'r') as file:
    csv_data = csv.DictReader(file)
    for row in csv_data:
        if row["urls_list"] is not None:

            try:
                urls_list = json.loads(row["urls_list"].replace("'", '"'))
                expanded_urls = [item.get("expanded_url") for item in urls_list]
                
                #new column creation containing the list of expandend urls
                row["expanded_url"] = expanded_urls
                tw_collection.insert_one(row)

            except (json.JSONDecodeError, TypeError):
                print(f"Error line {i} Url_list {row['urls_list']}")
        
        
        if (i % 10000) == 0:
            print(i)
        i += 1

# Youtube collection creation

# Reading CSV and collection creation

yt_collection = db1.yt_collection
with open('/Users/raffaelerusso/Documents/GitHub/Youtube-Video-Classification-on-Twitter/VideoClassification/y_train.csv', 'r') as file:
    csv_data = csv.DictReader(file)
    for row in csv_data:
       yt_collection.insert_one(row)

# Associate each yt video to a list of the referencing tweet ids

# Collect all id values from yt_collection
id_values = [doc['id'] for doc in yt_collection.find({}, {'id': 1})]

# Define the batch size
batch_size = 1000

# Split id_values into batches
id_batches = [id_values[i:i+batch_size] for i in range(0, len(id_values), batch_size)]

# Create a dictionary to store the matching tweet_ids for each id
id_tweet_ids = {id_val: set() for id_val in id_values}

# Iterate over each batch and collect the matching tweet_ids
for batch_index, id_batch in enumerate(id_batches):
    pattern = "|".join(map(re.escape, id_batch))

    # Find documents in tw_collection where expanded_url contains any of the id values in the batch
    results = tw_collection.find(
        {"expanded_url": {"$elemMatch": {"$regex": pattern}}},
        {"tweetid": 1, "expanded_url": 1}
    )
   
    # Update id_tweet_ids dictionary with the matching tweet_ids
    for doc in results:
        tweet_id = doc["tweetid"]
        expanded_url_list = doc["expanded_url"]
        for expanded_url in expanded_url_list:
            for id_val in id_batch:
                if id_val in expanded_url:
                    id_tweet_ids[id_val].add(tweet_id)

    # Print progress
    print(f"Processed batch {batch_index+1}/{len(id_batches)}")

# Update yt_collection with the tweet_ids for each id
for id_val, tweet_ids in id_tweet_ids.items():
    yt_collection.update_one(
        {"id": id_val},
        {"$set": {"tweet_ids": list(tweet_ids)}}
    )

# Print completion message
print("Matching tweet_ids updated in yt_collection.")

id_tweet_ids_list = {k: list(v) for k, v in id_tweet_ids.items()}

# Save id_tweet_ids to a file
with open("id_tweet_ids.json", "w") as file:
    json.dump(id_tweet_ids_list, file)

print("id_tweet_ids saved to file.")

# Creation of the new y_train CSV with the updated list of tweet ids 

path_labels = "./y_train.csv"
youtube_df = pd.read_csv(path_labels)

# Creation of the new dataframe
df = youtube_df.assign(tweet_ids=lambda x: x["id"].map(id_tweet_ids_list))

# Convert "not moderated" to 0 and "moderated" to 1
df["moderationStatus"] = df["moderationStatus"].replace({"not moderated": 0, "moderated": 1})

# Dropping column "lista_url"
df = df.drop("lista_url",axis=1)

# Save the updated DataFrame to a new CSV file
df.to_csv("updated_merged.csv", index=False)


print("Updated dataset saved to file.")

df.head()

# Saving the csv as a collection 

# Custom JSON encoder class to handle ObjectId serialization
class JSONEncoder(json.JSONEncoder):
    def default(self, o):
        if isinstance(o, ObjectId):
            return str(o)
        return super().default(o)

# Retrieve all documents from the collection
all_documents = list(yt_collection.find())

# Specify the path and filename for the output JSON file
output_file = "collection_data.json"

# Save the collection data to the JSON file
with open(output_file, "w") as file:
    json.dump(all_documents, file, cls=JSONEncoder)

print("Collection data saved as JSON.")
