==> Youtube-Video-Classification-on-Twitter/scripts/2-filtering_columns.ipynb <==
import pandas as pd 
import numpy as np 

# Twitter dataset reading and filtering

path_dataset = "df_youtube.csv"
tweets_df = pd.read_csv(path_dataset)
print(f"Dataset size: {tweets_df.shape}")

# Keeping only the tweets belonging to the types: reply, original, retweeted_tweet_without_comment and quoted tweet
tweets_filtered = tweets_df[(tweets_df["tweet_type"]=="reply") | (tweets_df["tweet_type"]=="original") | (tweets_df["tweet_type"]=="retweeted_tweet_without_comment") | (tweets_df["tweet_type"]=="quoted_tweet")]

# Dropping irrelevant columns
columns_to_keep = ['tweetid', 'userid', 'date', 'description', 'text', 'reply_statusid', 'tweet_type', 'friends_count', 'listed_count', 'followers_count', 'favourites_count', 'statuses_count', 'verified', 'hashtag', 'date_first_tweet', 'account_creation_date', 'rt_qtd_count', 'rt_rt_count', 'rt_reply_count', 'rt_fav_count', 'rt_tweetid', 'qtd_qtd_count', 'qtd_rt_count', 'qtd_reply_count', 'qtd_fav_count', 'qtd_tweetid']
tweets_filtered = tweets_df[(tweets_df["tweet_type"]=="reply") | (tweets_df["tweet_type"]=="original") | (tweets_df["tweet_type"]=="retweeted_tweet_without_comment") | (tweets_df["tweet_type"]=="quoted_tweet")][columns_to_keep]

tweets_filtered["tweet_type"].value_counts()

# Saving the filtered tweet dataset 

tweets_filtered.to_csv("tweets_filtered.csv",index=False)
