==> Youtube-Video-Classification-on-Twitter/scripts/2.2-baseline.ipynb <==
# Importing libraries

import pandas as pd 
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize
import string
import numpy as np
import xgboost as xgb
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score,f1_score,confusion_matrix,recall_score, precision_score
from sklearn.feature_extraction.text import CountVectorizer


nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')

# Reading youtube csv containing the field text

path = '/merged_yt_collection_text.csv'
df = pd.read_csv(path)

# Defining text preprocessing function

def preprocess_text(text):
    # Tokenize the text into individual words
    tokens = word_tokenize(text.lower())

    # Remove the punctuation
    table = str.maketrans('', '', string.punctuation)
    tokens = [token.translate(table) for token in tokens if token.isalpha()]

    # Remove the English stop words
    stop_words = set(stopwords.words('english'))
    tokens = [token for token in tokens if token not in stop_words]

    # Using a word Lemmatizer 
    lemmatizer = WordNetLemmatizer()
    tokens = [lemmatizer.lemmatize(token) for token in tokens]

    preprocessed_text = ' '.join(tokens)

    return preprocessed_text

# Creation of a new column with the preprocessed text

df['preprocessed_text'] = df['tweet_text'].apply(lambda x: preprocess_text(x))

# Saving the new csv with the preprocessed text

df.to_csv("merged_yt_collection_preprocessedtxt.csv",index=False)

# Check script mood.py

# Reading the csv with the preprocessed text


df = pd.read_csv("merged_yt_collection_preprocessedtxt.csv")

# Concat the new five mood columns to the dataframe

# Specify the file path
files = ['hate.txt', 'irony.txt', 'neutral.txt', 'negative.txt', 'offensive.txt']

# Read the text files into pandas Series
new_columns = []
for file in files:
    with open(file, 'r') as f:
        values = f.read().splitlines()
    col = pd.Series(values, name=file.split(".")[0])
    new_columns.append(col)

# Concatenate the existing DataFrame with the new mood columns
combined_data = pd.concat([df] + new_columns, axis=1)


combined_data.columns

# Extract input and target features

y = combined_data["moderationStatus"].to_numpy()
x = combined_data[["hate","irony","offensive","negative","neutral"]].to_numpy()

# Split data in train and test

X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.1, random_state=42,stratify=y)

# XGBoost

#Convert your data into DMatrix format (an optimized data structure for XGBoost)
dtrain = xgb.DMatrix(X_train, label=y_train)
dtest = xgb.DMatrix(X_test, label=y_test)

#Set the parameters for XGBoost
params = {
    'max_depth': 3,
    'eta': 0.1,
    'objective': 'binary:logistic',
    'eval_metric': 'logloss',
    'seed': 42
}

#Train the XGBoost model
num_rounds = 100
model = xgb.train(params, dtrain, num_rounds)

#Make predictions on the test set
y_pred = model.predict(dtest)
predictions = [round(value) for value in y_pred]

#Evaluate the model
accuracy = accuracy_score(y_test, predictions)
fscore = f1_score(y_test,predictions,average="macro")
recall = recall_score(y_test,predictions)
precision = precision_score(y_test,predictions)
conf_matr = confusion_matrix(y_test,predictions)
print("Accuracy: %.2f%%" % (accuracy * 100.0))
print("MACRO f1_Score: ",fscore)
print("recall: ",recall)
print("precision: ",precision)
print("confusion matrix: ",conf_matr)

# Random forest

#Split in training and test set
X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42)

# Create an instance of the Random Forest classifier
random_forest = RandomForestClassifier(n_estimators=100, random_state=42)  # Adjust the number of estimators as needed

# Fit the model to the training data
random_forest.fit(X_train, y_train)

# Make predictions on the test data
y_pred = random_forest.predict(X_test)

# Evaluate the model's performance
#Evaluate the model
accuracy = accuracy_score(y_test, y_pred)
fscore = f1_score(y_test,y_pred,average="macro")
recall = recall_score(y_test,y_pred)
precision = precision_score(y_test,y_pred)
conf_matr = confusion_matrix(y_test,y_pred)
print("Accuracy: %.2f%%" % (accuracy * 100.0))
print("macro f1_Score: ",fscore)
print("recall: ",recall)
print("precision: ",precision)
print("confusion matrix: ",conf_matr)

# SVM

# Split the data into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42)

# Create an instance of the SVM classifier
svm = SVC(kernel='sigmoid')  # You can choose different kernels, such as 'linear', 'rbf', or 'sigmoid'

# Fit the model to the training data
svm.fit(X_train, y_train)

# Make predictions on the test data
y_pred = svm.predict(X_test)

# Evaluate the model's performance
#Evaluate the model
accuracy = accuracy_score(y_test, y_pred)
fscore = f1_score(y_test,y_pred,average="macro")
recall = recall_score(y_test,y_pred)
precision = precision_score(y_test,y_pred)
conf_matr = confusion_matrix(y_test,y_pred)
print("accuracy: %.2f%%" % (accuracy * 100.0))
print("macro f1_Score: ",fscore)
print("recall: ",recall)
print("precision: ",precision)
print("confusion matrix: ",conf_matr)

# Logistic regression

# Split the data into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42)

# Create an instance of the SVM classifier
lr = LogisticRegression()  # You can choose different kernels, such as 'linear', 'rbf', or 'sigmoid'

# Fit the model to the training data
lr.fit(X_train, y_train)

# Make predictions on the test data
y_pred = lr.predict(X_test)

# Evaluate the model's performance
#Evaluate the model
accuracy = accuracy_score(y_test, y_pred)
fscore = f1_score(y_test,y_pred,average="macro")
recall = recall_score(y_test,y_pred)
precision = precision_score(y_test,y_pred)
conf_matr = confusion_matrix(y_test,y_pred)
print("Accuracy: %.2f%%" % (accuracy * 100.0))
print("macro f1_Score: ",fscore)
print("recall: ",recall)
print("precision: ",precision)
print("confusion matrix: ",conf_matr)

# Random Forest with the add of preprocessed text

text = combined_data["preprocessed_text"]

#text encoding
vect = CountVectorizer()
x_text = vect.fit_transform(text)

# Training and test set split
X_train,X_test,Y_Train,Y_test = train_test_split(x_text,y,test_size=0.2,random_state=42)

# Create an instance of the SVM classifier
model = RandomForestClassifier()  # You can choose different kernels, such as 'linear', 'rbf', or 'sigmoid'

# Fit the model to the training data
model.fit(X_train, y_train)

# Make predictions on the test data
y_pred = model.predict(X_test)

# Evaluate the model's performance
#Evaluate the model
accuracy = accuracy_score(y_test, y_pred)
fscore = f1_score(y_test,y_pred,average="macro")
recall = recall_score(y_test,y_pred)
precision = precision_score(y_test,y_pred)
conf_matr = confusion_matrix(y_test,y_pred)

print("Accuracy: %.2f%%" % (accuracy * 100.0))
print("f1_Score: ",fscore)
print("recall: ",recall)
print("precision: ",precision)
print("confusion matrix: ",conf_matr)
