==> Youtube-Video-Classification-on-Twitter/scripts/6-BERT.ipynb <==
# Import libraries

!pip install transformers

import tensorflow as tf
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.metrics import f1_score
import json
import torch
from transformers import BertTokenizer, BertForSequenceClassification, AdamW
from torch.utils.data import DataLoader

from google.colab import drive
drive.mount('/content/drive')

# Upload csv, extract the text input feature and the target variable y

df = pd.read_csv("/content/drive/MyDrive/BDE/BERT/merged_yt_collection_preprocessedtxt.csv")
text = df["preprocessed_text"].to_numpy()
y = df["moderationStatus"].to_numpy()

# Train and test split

test_size = 0.2
X_train,X_test,Y_train,Y_test = train_test_split(text,y,test_size=test_size,random_state=42,stratify = y)

# Compiling and training of the model

X_train_list = list(X_train)
Y_train_list = list(Y_train)

# Set device (GPU if available, otherwise CPU)
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

# Load pre-trained BERT model and tokenizer
model_name = 'bert-base-uncased'
model = BertForSequenceClassification.from_pretrained(model_name, num_labels=2).to(device)
tokenizer = BertTokenizer.from_pretrained(model_name)

# Load and preprocess the data
train_texts = X_train_list   # List of training texts
train_labels = Y_train_list  # List of corresponding training labels (0 or 1)

def preprocess(texts, labels):
    inputs = tokenizer(texts, padding=True, truncation=True, return_tensors='pt')
    inputs = {k: v.to(device) for k, v in inputs.items()}
    labels = torch.tensor(labels).to(device)
    return inputs, labels

train_inputs, train_labels = preprocess(train_texts, train_labels)
train_data = list(zip(train_inputs['input_ids'], train_inputs['attention_mask'], train_labels))
train_dataloader = DataLoader(train_data, batch_size=16, shuffle=True)

# Fine-tuning parameters
epochs = 3
learning_rate = 2e-5

# Set model to training mode
model.train()

# Initialize optimizer and loss function
optimizer = AdamW(model.parameters(), lr=learning_rate)
loss_fn = torch.nn.CrossEntropyLoss()

# Fine-tuning loop
for epoch in range(epochs):
    total_loss = 0

    for batch in train_dataloader:
        optimizer.zero_grad()
        input_ids, attention_mask, labels = [tensor.to(device) for tensor in batch]

        # Forward pass
        outputs = model(input_ids, attention_mask=attention_mask, labels=labels)
        loss = outputs.loss
        logits = outputs.logits

        # Backward pass
        loss.backward()
        optimizer.step()

        total_loss += loss.item()

    average_loss = total_loss / len(train_dataloader)
    print(f"Epoch {epoch+1}: Average Loss = {average_loss}")


# Save the fine-tuned model
model.save_pretrained('/content/drive/MyDrive/BDE/Models/fine-tuned-bert')

# Load the fine-tuned model
loaded_model = BertForSequenceClassification.from_pretrained('/content/drive/MyDrive/BDE/Models/fine-tuned-bert')

# Evaluating the model

input_inputs, _ = preprocess(list(X_test), [])  # Pass empty labels as we won't use them for inference
input_data = list(zip(input_inputs['input_ids'], input_inputs['attention_mask']))
input_dataloader = DataLoader(input_data, batch_size=16, shuffle=False)

loaded_model.eval()

input_predictions = []
loaded_model = loaded_model.to(device)
with torch.no_grad():
    for batch in input_dataloader:
        input_ids, attention_mask = batch
        inputs = {'input_ids': input_ids.to(device), 'attention_mask': attention_mask.to(device)}

        outputs = loaded_model(**inputs)
        logits = outputs.logits

        predicted_labels = torch.max(logits, dim=1).indices

        input_predictions.extend(predicted_labels.cpu().numpy())

input_predictions = np.array(input_predictions)

input_predictions


macro_f1_score = f1_score(Y_test,input_predictions,average="macro")
print(f"Macro f1-score: {macro_f1_score}")

# Save model

# Predict on test set

path_test = "/content/drive/MyDrive/BDE/Filtered_collections/y_test_collection.json"
csv_test_path  = "/content/drive/MyDrive/BDE/Predictions/y_test.csv"

with open(path_test, 'r') as test_file:
    test_json = json.load(test_file)

df_test= pd.read_csv(csv_test_path)

df_test_text = pd.read_csv("/content/drive/MyDrive/BDE/RNN/merged_yt_collection_test_preprocessedtxt.csv")
text_test = df_test_text["preprocessed_text"].to_numpy()

input_inputs, _ = preprocess(list(text_test), [])  # Pass empty labels as we won't use them for inference
input_data = list(zip(input_inputs['input_ids'], input_inputs['attention_mask']))
input_dataloader = DataLoader(input_data, batch_size=16, shuffle=False)

loaded_model.eval()

input_predictions_test = []
loaded_model = loaded_model.to(device)
with torch.no_grad():
    for batch in input_dataloader:
        input_ids, attention_mask = batch
        inputs = {'input_ids': input_ids.to(device), 'attention_mask': attention_mask.to(device)}

        outputs = loaded_model(**inputs)
        logits = outputs.logits

        predicted_labels = torch.max(logits, dim=1).indices

        input_predictions_test.extend(predicted_labels.cpu().numpy())

input_predictions_test = np.array(input_predictions_test)

df_test["moderationStatus"] =  np.where(input_predictions_test == 0, "not moderated", "moderated") 
df_test.to_csv(f"/content/drive/MyDrive/BDE/Predictions/BERT_model_{macro_f1_score:.10f}_y.csv",index=False)
