==> Youtube-Video-Classification-on-Twitter/scripts/6_test_rnn.ipynb <==
from pymongo import MongoClient
import json
import re
import csv
from bson import ObjectId
import ast
import string
import pandas as pd
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')

# Create a MongoClient object and specify the MongoDB connection URL
url = "mongodb://localhost:27017/"
client = MongoClient(url)

# Access a specific database
db = client["YT_db"]

# Send a ping to confirm a successful connection
try:
    client.admin.command('ping')
    print("Pinged your deployment. You successfully connected to MongoDB!")
except Exception as e:
    print(e)

# Reading twitter and the updated yt test collection

tw_collection = db.tw_collection

yt_test_collection = db.yt_test_collection


yt_test_collection.find_one()

# Retrieving tweets text and hashtags associated to each test yt video

def update_video_tweets(yt_collection, tw_collection):
    video_ids = yt_collection.find()  
    i = 0
    
    for video in video_ids:
        if i%10==0:
            print(i)
        i+=1
        video_id = video["id"] 
        
        #tweet_ids_str = video.get("tweet_ids", "")
        #tweet_ids = ast.literal_eval(tweet_ids_str)

        tweet_ids = video.get("tweet_ids", "")

        # Retrieve corresponding tweets from tw_collection
        tweets = tw_collection.find(
            {"tweetid": {"$in": tweet_ids}}, 
            {"text": 1, "hashtag": 1} 
        )

        # Collect text and hashtag for each tweet
        tweet_text = " "
        for tweet in tweets:             
            text = tweet.get("text") 
            if len(text) > 4:
                tweet_text += " " + text

            hashtag = tweet.get("hashtag")
            if len(hashtag) > 4:
                tweet_text += " " + hashtag

        # Update the video document with the tweet_text field
        yt_collection.update_one(
            {"id": video_id},
            {"$set": {"tweet_text": tweet_text}}
        )


update_video_tweets(yt_collection=yt_test_collection, tw_collection=tw_collection)

yt_test_collection.find_one()

# Saving as a json


# Custom JSON encoder class to handle ObjectId serialization
class JSONEncoder(json.JSONEncoder):
    def default(self, o):
        if isinstance(o, ObjectId):
            return str(o)
        return super().default(o)

# Retrieve all documents from the collection
all_documents = list(yt_test_collection.find())

# Specify the path and filename for the output JSON file
output_file = "yt_collection_test_text.json"

# Save the collection data to the JSON file
with open(output_file, "w") as file:
    json.dump(all_documents, file, cls=JSONEncoder,indent=4)

print("Collection data saved as JSON.")

# Saving as a csv

# Retrieve all documents from the collection
cursor = yt_test_collection.find()

# Convert the cursor to a DataFrame
df = pd.DataFrame(list(cursor))

# Specify the columns to include in the CSV file
columns = ['id', 'tweet_ids', 'tweet_text']

# Save the DataFrame as a CSV file
df.to_csv('merged_yt_collection_test_text.csv', columns=columns, index=False)

# Preprocess text

def preprocess_text(text):
    # Tokenize the text into individual words
    tokens = word_tokenize(text.lower())

    # Remove the punctuation
    table = str.maketrans('', '', string.punctuation)
    tokens = [token.translate(table) for token in tokens if token.isalpha()]

    # Remove the English stop words
    stop_words = set(stopwords.words('english'))
    tokens = [token for token in tokens if token not in stop_words]

    # Using a word Lemmatizer 
    lemmatizer = WordNetLemmatizer()
    tokens = [lemmatizer.lemmatize(token) for token in tokens]

    preprocessed_text = ' '.join(tokens)

    return preprocessed_text

df['preprocessed_text'] = df['tweet_text'].apply(lambda x: preprocess_text(x))

# Saving as a new csv

df.to_csv("merged_yt_collection_test_preprocessedtxt.csv",index=False)
